+++ app/utils/map.ts
@@ -1,0 +1,30 @@
+import { TransformedSearchHit } from "models";
+
+export const groupHitsByLocation = (hits: TransformedSearchHit[]) => {
+  return hits.reduce((acc, hit) => {
+    hit.locations.forEach((location) => {
+      const key = `${location.lat}-${location.long}`;
+      if (!acc[key]) {
+        acc[key] = [];
+      }
+      acc[key].push({ hit, location });
+    });
+    return acc;
+  }, {} as Record<string, { hit: TransformedSearchHit; location: { id: string; lat: string; long: string; label: string } }[]>);
+};
+
+export const computeGridOffset = (
+  index: number,
+  total: number,
+  epicenterLat: number,
+  epicenterLng: number,
+  spacing: number = 0.00004
+): { offsetLat: number; offsetLng: number } => {
+  const cols = Math.ceil(Math.sqrt(total));
+  const rows = Math.ceil(total / cols);
+  const row = Math.floor(index / cols);
+  const col = index % cols;
+  const offsetLat = epicenterLat + (row - (rows - 1) / 2) * spacing;
+  const offsetLng = epicenterLng + (col - (cols - 1) / 2) * spacing;
+  return { offsetLat, offsetLng };
+};
